+++ gastropaysdk/src/main/java/com/inventiv/gastropaysdk/data/request/InvoiceSendRequest.kt
@@ -1,0 +1,5 @@
+package com.inventiv.gastropaysdk.data.request
+
+internal data class InvoiceSendRequest(
+    val walletTransactionId: Int
+)