+++ gastropaysdk/src/main/java/com/inventiv/gastropaysdk/ui/wallet/TransactionItemViewHolder.kt
@@ -1,10 +1,12 @@
 package com.inventiv.gastropaysdk.ui.wallet
 
 import android.graphics.Color
+import androidx.core.content.ContextCompat
 import androidx.recyclerview.widget.RecyclerView
 import com.inventiv.gastropaysdk.R
 import com.inventiv.gastropaysdk.data.TransactionModel
 import com.inventiv.gastropaysdk.databinding.ItemWalletTransactionGastropaySdkBinding
+import com.inventiv.gastropaysdk.utils.TransactionType
 import com.inventiv.gastropaysdk.utils.blankj.utilcode.util.StringUtils
 import com.inventiv.gastropaysdk.utils.formatDate
 
@@ -24,14 +26,35 @@ internal class TransactionItemViewHolder(
             valueTextView.text = transaction.price
             dateTextView.text = transaction.date.formatDate("dd.MM.yyyy - HH:mm")
 
-            if (transaction.isEarn) {
-                typeTextView.text =
-                    StringUtils.getString(R.string.wallet_earn_gastropay_sdk)
-                typeTextView.setTextColor(Color.parseColor("#2ECC71"))
-            } else {
-                typeTextView.text =
-                    StringUtils.getString(R.string.wallet_spend_gastropay_sdk)
-                typeTextView.setTextColor(Color.parseColor("#FD3850"))
+            when (transaction.transactionType) {
+                TransactionType.DEPOSIT -> {
+                    typeTextView.text = StringUtils.getString(R.string.wallet_earn_gastropay_sdk)
+                    typeTextView.setTextColor(
+                        ContextCompat.getColor(
+                            itemView.context,
+                            R.color.shamrock_gastropay_sdk
+                        )
+                    )
+                }
+                TransactionType.WITHDRAW -> {
+                    typeTextView.text = StringUtils.getString(R.string.wallet_spend_gastropay_sdk)
+                    typeTextView.setTextColor(
+                        ContextCompat.getColor(
+                            itemView.context,
+                            R.color.reddish_orange_gastropay_sdk
+                        )
+                    )
+                }
+                TransactionType.CANCEL_DEPOSIT,
+                TransactionType.CANCEL_WITHDRAW -> {
+                    typeTextView.text = StringUtils.getString(R.string.wallet_cancel_gastropay_sdk)
+                    typeTextView.setTextColor(
+                        ContextCompat.getColor(
+                            binding.root.context,
+                            R.color.suva_grey_gastropay_sdk
+                        )
+                    )
+                }
             }
 
             if (position % 2 == 0) {